Serialize the text field and the type field of AutoTextContent

+++ engine/org.eclipse.birt.report.engine/src/org/eclipse/birt/report/engine/content/IAutoTextContent.java
@@ -3,10 +3,10 @@ package org.eclipse.birt.report.engine.content;
 
 public interface IAutoTextContent extends ITextContent
 {
-	public static final byte TOTAL_PAGE = 0;
-	public static final byte PAGE_NUMBER = 1;
+	public static final int TOTAL_PAGE = 0;
+	public static final int PAGE_NUMBER = 1;
 	
-	void setType ( byte type );
+	void setType ( int type );
 	
-	byte getType ( );
+	int getType ( );
 }
